Create input module and move relevant files there
