+++ examples/swash_render/src/main.rs
@@ -12,8 +12,8 @@
 
 use image::codecs::png::PngEncoder;
 use image::{self, Pixel, Rgba, RgbaImage};
+use parley::inputs::style::{FontStack, FontWeight, StyleProperty, TextStyle};
 use parley::layout::{Alignment, Glyph, GlyphRun, Layout, PositionedLayoutItem};
-use parley::style::{FontStack, FontWeight, StyleProperty, TextStyle};
 use parley::{FontContext, InlineBox, LayoutContext};
 use std::fs::File;
 use swash::scale::image::Content;
