+++ parley/src/inputs/mod.rs
@@ -1,0 +1,13 @@
+// Copyright 2024 the Parley Authors
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+// TODO - Remove pub(crate)
+
+pub(crate) mod builder;
+pub(crate) mod context;
+pub(crate) mod font;
+pub(crate) mod line_metrics;
+
+pub mod style;
+
+pub use line_metrics::*;
