+++ parley/src/layout/data.rs
@@ -3,7 +3,7 @@
 
 use crate::inline_box::InlineBox;
 use crate::layout::{Alignment, Glyph, LineMetrics, RunMetrics, Style};
-use crate::style::Brush;
+use crate::inputs::style::Brush;
 use crate::util::nearly_zero;
 use crate::Font;
 use core::ops::Range;
